+++ mio/src/sys/unix/uds/mod.rs
@@ -1,66 +1,5 @@
-use std::cmp::Ordering;
-use std::os::unix::ffi::OsStrExt;
-use std::path::Path;
-use std::{io, mem};
-
-mod datagram;
-pub use self::datagram::UnixDatagram;
-
-mod listener;
-pub use self::listener::{SocketAddr, UnixListener};
-
-mod stream;
-pub use self::stream::UnixStream;
-
-pub fn socket_addr(path: &Path) -> io::Result<(libc::sockaddr_un, libc::socklen_t)> {
-    let sockaddr = mem::MaybeUninit::<libc::sockaddr_un>::zeroed();
-
-    // This is safe to assume because a `libc::sockaddr_un` filled with `0`
-    // bytes is properly initialized.
-    //
-    // `0` is a valid value for `sockaddr_un::sun_family`; it is
-    // `libc::AF_UNSPEC`.
-    //
-    // `[0; 108]` is a valid value for `sockaddr_un::sun_path`; it begins an
-    // abstract path.
-    let mut sockaddr = unsafe { sockaddr.assume_init() };
-
-    sockaddr.sun_family = libc::AF_UNIX as libc::sa_family_t;
-
-    let bytes = path.as_os_str().as_bytes();
-    match (bytes.get(0), bytes.len().cmp(&sockaddr.sun_path.len())) {
-        // Abstract paths don't need a null terminator
-        (Some(&0), Ordering::Greater) => {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "path must be no longer than libc::sockaddr_un.sun_path",
-            ));
-        }
-        (_, Ordering::Greater) | (_, Ordering::Equal) => {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "path must be shorter than libc::sockaddr_un.sun_path",
-            ));
-        }
-        _ => {}
-    }
-
-    for (dst, src) in sockaddr.sun_path.iter_mut().zip(bytes.iter()) {
-        *dst = *src as libc::c_char;
-    }
-
-    let offset = path_offset(&sockaddr);
-    let mut socklen = offset + bytes.len();
-
-    match bytes.get(0) {
-        // The struct has already been zeroes so the null byte for pathname
-        // addresses is already there.
-        Some(&0) | None => {}
-        Some(_) => socklen += 1,
-    }
-
-    Ok((sockaddr, socklen as libc::socklen_t))
-}
+mod socketaddr;
+pub use self::socketaddr::SocketAddr;
 
 /// Get the `sun_path` field offset of `sockaddr_un` for the target OS.
 ///
@@ -74,38 +13,104 @@ pub fn path_offset(sockaddr: &libc::sockaddr_un) -> usize {
     path - base
 }
 
-#[cfg(test)]
-mod tests {
-    use super::{path_offset, socket_addr};
+cfg_os_poll! {
+    use std::cmp::Ordering;
+    use std::os::unix::ffi::OsStrExt;
     use std::path::Path;
-    use std::str;
+    use std::{io, mem};
 
-    #[test]
-    fn pathname_address() {
-        const PATH: &str = "./foo/bar.txt";
-        const PATH_LEN: usize = 13;
+    mod datagram;
+    pub use self::datagram::UnixDatagram;
 
-        // Pathname addresses do have a null terminator, so `socklen` is
-        // expected to be `PATH_LEN` + `offset` + 1.
-        let path = Path::new(PATH);
-        let (sockaddr, actual) = socket_addr(path).unwrap();
+    mod listener;
+    pub use self::listener::UnixListener;
+
+    mod stream;
+    pub use self::stream::UnixStream;
+
+    pub fn socket_addr(path: &Path) -> io::Result<(libc::sockaddr_un, libc::socklen_t)> {
+        let sockaddr = mem::MaybeUninit::<libc::sockaddr_un>::zeroed();
+
+        // This is safe to assume because a `libc::sockaddr_un` filled with `0`
+        // bytes is properly initialized.
+        //
+        // `0` is a valid value for `sockaddr_un::sun_family`; it is
+        // `libc::AF_UNSPEC`.
+        //
+        // `[0; 108]` is a valid value for `sockaddr_un::sun_path`; it begins an
+        // abstract path.
+        let mut sockaddr = unsafe { sockaddr.assume_init() };
+
+        sockaddr.sun_family = libc::AF_UNIX as libc::sa_family_t;
+
+        let bytes = path.as_os_str().as_bytes();
+        match (bytes.get(0), bytes.len().cmp(&sockaddr.sun_path.len())) {
+            // Abstract paths don't need a null terminator
+            (Some(&0), Ordering::Greater) => {
+                return Err(io::Error::new(
+                    io::ErrorKind::InvalidInput,
+                    "path must be no longer than libc::sockaddr_un.sun_path",
+                ));
+            }
+            (_, Ordering::Greater) | (_, Ordering::Equal) => {
+                return Err(io::Error::new(
+                    io::ErrorKind::InvalidInput,
+                    "path must be shorter than libc::sockaddr_un.sun_path",
+                ));
+            }
+            _ => {}
+        }
+
+        for (dst, src) in sockaddr.sun_path.iter_mut().zip(bytes.iter()) {
+            *dst = *src as libc::c_char;
+        }
+
         let offset = path_offset(&sockaddr);
-        let expected = PATH_LEN + offset + 1;
-        assert_eq!(expected as libc::socklen_t, actual)
+        let mut socklen = offset + bytes.len();
+
+        match bytes.get(0) {
+            // The struct has already been zeroes so the null byte for pathname
+            // addresses is already there.
+            Some(&0) | None => {}
+            Some(_) => socklen += 1,
+        }
+
+        Ok((sockaddr, socklen as libc::socklen_t))
     }
 
-    #[test]
-    fn abstract_address() {
-        const PATH: &[u8] = &[0, 116, 111, 107, 105, 111];
-        const PATH_LEN: usize = 6;
+    #[cfg(test)]
+    mod tests {
+        use super::{path_offset, socket_addr};
+        use std::path::Path;
+        use std::str;
 
-        // Abstract addresses do not have a null terminator, so `socklen` is
-        // expected to be `PATH_LEN` + `offset`.
-        let abstract_path = str::from_utf8(PATH).unwrap();
-        let path = Path::new(abstract_path);
-        let (sockaddr, actual) = socket_addr(path).unwrap();
-        let offset = path_offset(&sockaddr);
-        let expected = PATH_LEN + offset;
-        assert_eq!(expected as libc::socklen_t, actual)
+        #[test]
+        fn pathname_address() {
+            const PATH: &str = "./foo/bar.txt";
+            const PATH_LEN: usize = 13;
+
+            // Pathname addresses do have a null terminator, so `socklen` is
+            // expected to be `PATH_LEN` + `offset` + 1.
+            let path = Path::new(PATH);
+            let (sockaddr, actual) = socket_addr(path).unwrap();
+            let offset = path_offset(&sockaddr);
+            let expected = PATH_LEN + offset + 1;
+            assert_eq!(expected as libc::socklen_t, actual)
+        }
+
+        #[test]
+        fn abstract_address() {
+            const PATH: &[u8] = &[0, 116, 111, 107, 105, 111];
+            const PATH_LEN: usize = 6;
+
+            // Abstract addresses do not have a null terminator, so `socklen` is
+            // expected to be `PATH_LEN` + `offset`.
+            let abstract_path = str::from_utf8(PATH).unwrap();
+            let path = Path::new(abstract_path);
+            let (sockaddr, actual) = socket_addr(path).unwrap();
+            let offset = path_offset(&sockaddr);
+            let expected = PATH_LEN + offset;
+            assert_eq!(expected as libc::socklen_t, actual)
+        }
     }
 }
